chore(new-ui): wip
